+++ src/main/java/com/nhnacademy/ssacthree_shop_api/bookset/book/service/impl/BookCommonServiceImpl.java
@@ -214,13 +214,13 @@ public class BookCommonServiceImpl implements BookCommonService {
     }
 
     @Override
-    public Boolean deleteBookLike(Long bookId, Long customerId) {
+    public BookLikeResponse deleteBookLike(Long bookId, Long customerId) {
 
         BookLikeId bookLikeId = new BookLikeId(bookId, customerId);
 
         if (bookLikeRepository.existsById(bookLikeId)) {
             bookLikeRepository.deleteById(bookLikeId);
-            return true;
+            return new BookLikeResponse(bookId, bookRepository.findBookLikeByBookId(bookId));
         } else {
             throw new NotFoundException("해당 좋아요 기록이 존재하지 않습니다.");
         }
